+++ lib/CMakeLists.txt
@@ -17,6 +17,7 @@ add_subdirectory_ifdef(CONFIG_MODEM_INFO modem_info)
 add_subdirectory_ifdef(CONFIG_PDN_MANAGEMENT pdn_management)
 add_subdirectory_ifdef(CONFIG_MULTITHREADING_LOCK multithreading_lock)
 add_subdirectory_ifdef(CONFIG_ADP536X adp536x)
+add_subdirectory_ifdef(CONFIG_ST25R3911B_LIB st25r3911b)
 add_subdirectory_ifdef(CONFIG_FPROTECT fprotect)
 add_subdirectory(flash_patch)
 add_subdirectory_ifdef(CONFIG_RESET_ON_FATAL_ERROR fatal_error)
