+++ lib/Kconfig
@@ -10,6 +10,7 @@ rsource "bin/Kconfig"
 rsource "bsdlib/Kconfig"
 rsource "adp536x/Kconfig"
 rsource "at_cmd/Kconfig"
+rsource "st25r3911b/Kconfig"
 rsource "flash_patch/Kconfig"
 rsource "lte_link_control/Kconfig"
 rsource "fprotect/Kconfig"
